// src/vendor/interactive-backgrounds/DNASparkBackground.d.ts
import React from 'react';
interface DNASparkBackgroundProps {
    sparkColor?: string;
    strandColor?: string;
    className?: string;
    particleColor?: string;
    connectionColor?: string;
    rippleColor?: string;
    color?: string;
    constfill?: string;
    text?: string;
}
declare const DNASparkBackground: React.FC<DNASparkBackgroundProps>;
export default DNASparkBackground;