// src/vendor/interactive-backgrounds/DNASparkBackground3D.d.ts
import React from 'react';
interface DNASparkBackground3DProps {
    sparkColor?: string;
    strandColor?: string;
    className?: string;
    particleColor?: string;
    connectionColor?: string;
    rippleColor?: string;
    color?: string;
    constfill?: string;
    text?: string;
    basePairCount?: number;
    helixRadius?: number;
    helixHeight?: number;
    basePairSpacing?: number;
    backboneSegments?: number;
    backboneRadius?: number;
    baseSphereRadius?: number;
    baseSphereSegments?: number;
    particleCount?: number;
    particleSize?: number;
    particleSpeedMin?: number;
    particleSpeedMax?: number;
    glowRadius?: number;
    glowIntensityMultiplier?: number;
    cameraOrbitRadius?: number;
    cameraOrbitSpeed?: number;
    dnaRotationSpeed?: number;
    particleOrbitalAmplitude?: number;
    rippleGrowthRate?: number;
    rippleFadeRate?: number;
}
declare const DNASparkBackground3D: React.FC<DNASparkBackground3DProps>;
export default DNASparkBackground3D;